requirements.txt
openai
ijson
PyYAML
tiktoken
tqdm 